Conflicts:
	src/org/wordpress/android/ui/comments/CommentsActivity.java
	src/org/wordpress/android/ui/reader/ReaderPostDetailFragment.java

+++ src/org/wordpress/android/ui/comments/CommentDetailFragment.java
@@ -105,7 +105,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
     /*
      * used when called from comment list
      */
-    protected static CommentDetailFragment newInstance(int localBlogId, int commentId) {
+    protected static CommentDetailFragment newInstance(int localBlogId, long commentId) {
         CommentDetailFragment fragment = new CommentDetailFragment();
         fragment.setComment(localBlogId, commentId);
         return fragment;
@@ -194,7 +194,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
         return view;
     }
 
-    protected void setComment(int localBlogId, int commentId) {
+    protected void setComment(int localBlogId, long commentId) {
         setComment(localBlogId, CommentTable.getComment(localBlogId, commentId));
     }
 
@@ -264,7 +264,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
         return (mComment != null);
     }
 
-    private int getCommentId() {
+    private long getCommentId() {
         return (mComment != null ? mComment.commentID : 0);
     }
 
@@ -408,7 +408,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
      * ensure the post associated with this comment is available to the reader and show its
      * title above the comment
      */
-    private void showPostTitle(final int blogId, final int postId) {
+    private void showPostTitle(final int blogId, final long postId) {
         if (!hasActivity())
             return;
